Endpoints e testes funcionando

+++ src/main/java/caixaverso/application/dto/TelemetriaItemDTO.java
@@ -3,4 +3,4 @@ package caixaverso.application.dto;
 public record TelemetriaItemDTO(
         String nome,
         Long quantidadeChamadas,
-        Long mediaTempoRespostaMs) {}
+        Double mediaTempoRespostaMs) {}
